great symbol migration plus fixes
