+++ main.js
@@ -4,7 +4,6 @@ import { Catalog } from './catalog.js';
 import { Board } from './board.js';
 import { Inventory } from './inventory.js';
 import { Shop } from './shop.js';
-``;
 import { Game } from './game.js';
 import { Progression } from './progression.js';
 
@@ -30,7 +29,9 @@ export const loadSettings = async (settings = GameSettings.instance()) => {
       const symbol = game.catalog.symbol(emoji);
       if (symbol) {
         const interactiveDescription = Util.createInteractiveDescription(
-          symbol.descriptionLong(), /*emoji=*/symbol.emoji());
+          symbol.descriptionLong(),
+          /*emoji=*/ symbol.emoji()
+        );
         Util.drawText(game.info, interactiveDescription, true);
       }
     }
@@ -53,6 +54,10 @@ loadSettings(PROGRESSION.levelData[PROGRESSION.activeLevel]);
 
 ///// TEST RELATED CODE BELOW //////
 
+class SimBoard extends Board {
+  redrawCell(_, __, ___) {}
+}
+
 class AutoGame {
   constructor(gameSettings, catalog, buyAlways, buyOnce, buyRandom) {
     this.gameSettings = gameSettings;
@@ -62,7 +67,7 @@ class AutoGame {
       this.catalog.symbolsFromString(this.gameSettings.startingSet)
     );
     this.inventory.update();
-    this.board = new Board(this.gameSettings, this.catalog);
+    this.board = new SimBoard(this.gameSettings, this.catalog);
     this.shop = new Shop(this.catalog);
     this.totalTurns = 0;
     this.buyAlways = new Set(buyAlways);
