+++ shop.js
@@ -30,14 +30,18 @@ export class Shop {
       symbolDiv.innerHTML = symbol.emoji();
       symbolDiv.addEventListener('click', () => {
         const interactiveDescription = Util.createInteractiveDescription(
-          symbol.descriptionLong(), /*emoji=*/symbol.emoji());
+          symbol.descriptionLong(),
+          /*emoji=*/ symbol.emoji()
+        );
         Util.drawText(game.info, interactiveDescription, true);
       });
       shopItemDiv.appendChild(symbolDiv);
 
       const descriptionDiv = document.createElement('div');
       descriptionDiv.classList.add('description');
-      descriptionDiv.innerHTML = Util.createInteractiveDescription(symbol.description());
+      descriptionDiv.innerHTML = Util.createInteractiveDescription(
+        symbol.description()
+      );
       shopItemDiv.appendChild(descriptionDiv);
 
       const costDiv = document.createElement('div');
