+++ symbols/tutorial.js
@@ -1,0 +1,17 @@
+// This symbol file is a bit special. Rather than defining any symbols,
+// it demonstrates how a custom set can be re-exported for use by GameSettings
+
+export { Balloon } from './things.js';
+export { Bug, Dragon } from './animals.js';
+export {
+  Butter,
+  Corn,
+  Popcorn,
+  Cherry,
+  Mango,
+  Pineapple,
+  Tree,
+} from './food.js';
+export { Clover } from './advanced.js';
+export { Coin, Bank, MoneyBag, Briefcase } from './money.js';
+export { Diamond } from './rocks.js';
